Do not try to download Chromium by default
This happens because the wordpress/scripts package depends on Puppeteer, which results in attempts to download Chromium when running `npm install`.

Once we use Puppeteer for end-to-end tests that can be run on Travis, we can enable the download again just for that stage.

Related: https://github.com/WordPress/gutenberg/pull/15712