First draft of new story editor
This is only a first draft. It's not using any Gutenberg block functionality at all, but provides a framework for how to layout and organize the components.

Remember to `npm i` before trying to compile. Also not that this completely breaks actual story editing in every possible way.

How do you get rid of that annoying footer with the "thanks" and "version" by the way?