Merged the ideas of edit-mode and moveable
This was quite a complex merge. How do we both allow moveable to detect drags while the element itself must detect clicks? It does this by actually detecting clicks on the fake selection box and forwarding these clicks to the selected element. By far is this rigid, but it "works" in the basic case.

This also includes some updates to the moveable layout - including a much-needed update of the moveable rectangle when the selection changes.