Clean-up and proper edit mode
This commit cleans up some of the structure as well as correctly uses edit mode as in the UX.

There's still the issue of clicking the background which will both exit edit mode and unselect the element, as the default action on the background is that latter action. I don't know a good fix for that right now.