Import React dependencies directly from react and react-dom
This does NOT remove the dependency on the `@wordpress/element` package and does NOT affect bundle size, as React is still an external in the eyes of webpack.

However, this change makes it more clear what exactly we need from the package. This way it’s easier to eventually drop it in the future if deemed unnecessary.

Nevertheless, we might still need `@wordpress/element` specifically in the future for things like `__experimentalCreateInterpolateElement()` or others.