Reduce amount of  WP dependencies (#326)
* Remove dependency on `@wordpress/dom-ready` package
* Remove `@wordpress/url` package in favor of own custom `addQueryArgs` util
* Replace `@wordpress/a11y` with custom simple implementation
* Add DOM Testing Library utils and ESLint config
* Import React dependencies directly from `react` and `react-dom` instead of `@wordpress/element`

Note: this does NOT remove the dependency on the `@wordpress/element` package and does NOT affect bundle size, as React is still an external in the eyes of `webpack`.

However, this change makes it more clear what exactly we need from the package. This way it’s easier to eventually drop it in the future if deemed unnecessary.

Nevertheless, we might still need `@wordpress/element` specifically in the future for things like `__experimentalCreateInterpolateElement()` or others.