Changed exit to be on pointerdown, not click
This changes how the mouse-based exit detection works for both edit mode and color picker.

If you _mouse-down_ inside the component, but subsequently _mouse-up_ outside the component, the component should not treat this as an exit signal.

This is done by changed the listener to be a `pointerdown` listener rather than a `click` listener in both instances.

This might have unexpected consequences, so heavy testing is required.