Change the way videoAccessibility.js resolves an empty string in the text input.
Currently we default to the resource if the poster/title/alt strings are falsey. This is obviously a problem for an empty string (which is a valid value for the title and alt fields). To fix this, fallback to the resource only for undefined (which is the value that is returned when no changes have been made via the inspector).

Additionally, when calling pushUpdate default to null if it's an empty string. Unfortunately having an empty string will result in incorrect logic in updateProperties.js as it ignores the update when in commit mode.