Copy properties from original when temp replacing
The properties flip, scale, focusX and focusY will be copied from the original element when previewing. Because flip is handled by the mask, the replacement element has to be outside the mask and come with its own mask.

Finally this also uses `combineElements` when setting as background, to unify the way background is set whether you drag or use the button to do it.