Added keyboard shortcuts to tab views
This required allowing some digit keys to force display of keyboard outlines - it seems to work okay.

This also required to allow a key listener to abort on empty shortcut, as tab views could be without a shortcut (though we do not have that currently), as hooks cannot be conditional.