Use pull_request_target event for workflows
This behaves in an almost identical way to the pull_request event with the same set of filters and payload. However, instead of running against the workflow and code from the merge commit, the event runs against the workflow and code from the base of the pull request.

This is needed for PRs from repository forks, and makes it easier for non-members to contribute to the plugin.

See https://github.blog/2020-08-03-github-actions-improvements-for-fork-and-pull-request-workflows/ for details