Fixed tooltip tails (#6416)
top and bottom tail tooltips were off by 1 in app usage but not storybook - the left and right tails were already updated with - 1px but storybook made it seem that it was only an issue with l/r - this is because the SVG has 1px space below it. Also updating span to use forwardAs so we get font back.