Updated to new tooltips (#6411)
* Added new arrows using design system

* Using existing deepmerge lib, but not actively adding to package.json

This is only temporary until we come up with the proper solution in #6209

* Added new arrow buttons and spacing to carousel

* Changed buttons to circles because I think that is the intent?

* Updated to new icon buttons in the page menu

They do not look very good though. Will have to ask about that.

* Added page menu icons

Had to change some things around the page menu layout as well, but some of those overlaps with zoom PR i #4022, so will have some conflicts there.

* Generalized common props

* Removed unused flag

* Added new button pixels constants

* Employed new button pixel constants

* Added one more case of button pixel usage

* Changed all buttons to square (never circles!)

* Removed irrelevant css property

* Much better grid structure!

* Moved dashboard specific icons out of design system

* Replaced all icons with new ones

This has many consequences as the new icons are differently boxed compared to the old ones. I will try to fix in subsequent commits.

* Fixed icon styling

* Added small icons, fixed dropdown chevron

* Fixed cross icon in snackbar

* Changed all icon colors to dynamic and added storybook demo

* Fixed inspector icon usage and storybook

* Fixed plus icon in color picker

* Fixed xs chevron icons after alignment

* Fixed plus icons in color preset empty panel

* Fixed icons in preset panel header

* Swapped two icons

* Fixed design in help center including new icons

* Reversed help center chevron direction

* Renamed icon

* Updated instructions

* Made some icon adjustments

Make checkmark bigger in design system (used on dashboard/settings and telemetry banner), make toggle checkmark bigger (only exists in storybook), tweak prepublish alert icon that shows up on publisher or update button to be better size, move space for help center toggle to be padding instead of a div. (#6319)

* Edited icon instructions

* Fixed page menu sizing and structure

* Refactored page menu

* Fixed page count spacing

* Updated page count to reflect recent UX review

* Fixed some icon colors

* Fixed some icon colors

* Fixed icon size in storybook

* Renamed some icons

* Fixed position and css for ds toggle checkmark

* Scaled page layouts icon back up

* Updated styling of + in color picker text

* Updated help center toggle button layout

* Exit debug

* Used existed typography

* Updated to new tooltips

* Added missing tail to metabox toggle

* Fixed tooltip tails (#6416)

top and bottom tail tooltips were off by 1 in app usage but not storybook - the left and right tails were already updated with - 1px but storybook made it seem that it was only an issue with l/r - this is because the SVG has 1px space below it. Also updating span to use forwardAs so we get font back.

Co-authored-by: Brittany Feenstra <brittany.feenstra@formidable.com>