KSES: Improve compatibility for WP 5.4.1+
By leveraging a new param that was added to the `wp_insert_post_data` filter in WordPRess 5.4.1, we can better target posts of the web-story post type with our custom KSES logic.

The existing behavior is still in place for WP < 5.4.1, but should be removed once we drop support for those versions.

Originally discussed at google/web-stories-wp#6077