Media: Don't force "Upload files" as default tab
This prevents a race condition where the tab is switched after uploading, causing a force-refresh of the items in the "Media Library", causing the just uploaded item to be dropped from the list again.

Updates e2e tests accordingly to first switch to the "Upload files" tab before uploading files.